# backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FORJA API",
    "version": "0.1.0",
    "description": "Generado desde los schemas zod del backend (backend/src/openapi). No editar a mano — correr `npm run openapi:generate` después de cambiar una ruta."
  },
  "servers": [
    {
      "url": "http://localhost:4000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Token de sesión emitido por /auth/login o /auth/register."
      },
      "clientIdLegacy": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Client-Id",
        "description": "Modo legacy sin cuenta: identifica la instalación. Se ignora si viene Authorization Bearer válido."
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "nullable": true
          }
        }
      },
...
      "NewRecord": {
        "type": "object",
        "properties": {
          "exercise_name": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "weight",
              "volume"
            ]
          }
        },
        "required": [
          "exercise_name",
          "type"
        ]
      },
      "SessionResult": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "date": {
            "type": "string"
          },
          "workout_type_id": {
            "type": "string",
            "nullable": true
          },
          "custom_routine_id": {
            "type": "number",
            "nullable": true
          },
          "new_records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NewRecord"
            }
          }
        },
        "required": [
          "id",
          "date",
          "new_records"
        ]
      },
      "SessionSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "date": {
            "type": "string"
          },
          "workout_type_id": {
            "type": "string",
            "nullable": true
          },
          "custom_routine_id": {
            "type": "number",
            "nullable": true
          },
          "workout_label": {
            "type": "string"
          },
          "rpe": {
            "type": "number",
            "nullable": true
          }
        },
        "required": [
          "id",
          "date",
          "workout_type_id",
          "custom_routine_id",
          "workout_label",
          "rpe"
        ]
      },
      "SessionExerciseSet": {
        "type": "object",
        "properties": {
          "exercise_name": {
            "type": "string"
          },
          "weight_kg": {
            "type": "number"
          },
          "reps": {
            "type": "number"
          },
          "set_number": {
            "type": "number"
          }
        },
        "required": [
          "exercise_name",
          "weight_kg",
          "reps",
          "set_number"
        ]
      },
      "SessionDetail": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SessionSummary"
          },
          {
            "type": "object",
            "properties": {
              "tonelaje_total": {
                "type": "number"
              },
              "peso_medio": {
                "type": "number",
                "nullable": true
              },
              "intensidad_promedio_pct": {
                "type": "number",
                "nullable": true
              },
              "started_at": {
                "type": "string",
                "nullable": true
              },
              "ended_at": {
                "type": "string",
                "nullable": true
              },
              "indice_hipertrofia": {
                "type": "number",
                "nullable": true
              },
              "coeficiente_hipertrofia": {
                "type": "number",
                "nullable": true
              },
              "exercises": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/SessionExerciseSet"
                }
              }
            },
            "required": [
              "tonelaje_total",
              "peso_medio",
              "intensidad_promedio_pct",
              "started_at",
              "ended_at",
              "indice_hipertrofia",
              "coeficiente_hipertrofia",
              "exercises"
            ]
          }
        ]
      },
...
      "CardioSession": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "date": {
            "type": "string"
          },
          "activity_type": {
            "type": "string",
            "enum": [
              "cardio",
              "tecnico_tactico",
              "otro"
            ]
          },
          "duration_min": {
            "type": "number"
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "date",
          "activity_type",
          "duration_min",
          "notes"
        ]
      }
    },
...
  },
  "paths": {
...
    "/api/sessions": {
      "post": {
        "tags": [
          "sessions"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "clientIdLegacy": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "date": {
                    "type": "string",
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                  },
                  "workout_type_id": {
                    "type": "string",
                    "minLength": 1
                  },
                  "custom_routine_id": {
                    "type": "integer",
                    "minimum": 0,
                    "exclusiveMinimum": true
                  },
                  "exercises": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "exercise_name": {
                          "type": "string",
                          "minLength": 1
                        },
                        "sets": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "weight_kg": {
                                "type": "number",
                                "minimum": 0
                              },
                              "reps": {
                                "type": "integer",
                                "minimum": 0,
                                "exclusiveMinimum": true
                              },
                              "is_warmup": {
                                "type": "boolean",
                                "default": false
                              },
                              "rir": {
                                "type": "integer",
                                "minimum": 0,
                                "maximum": 10
                              }
                            },
                            "required": [
                              "weight_kg",
                              "reps"
                            ]
                          },
                          "minItems": 1
                        }
                      },
                      "required": [
                        "exercise_name",
                        "sets"
                      ]
                    },
                    "minItems": 1
                  },
                  "rpe": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 10
                  },
                  "started_at": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "ended_at": {
                    "type": "string",
                    "format": "date-time"
                  }
                },
                "required": [
                  "date",
                  "exercises"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Sesión guardada, con PRs nuevos si los hay",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResult"
                }
              }
            }
          },
          "400": {
            "description": "Error de validación",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Falta identidad (token inválido o X-Client-Id ausente)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Tipo/rutina no encontrados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "sessions"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "clientIdLegacy": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "from",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "to",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "as_athlete_id",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Calendario de sesiones",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SessionSummary"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error de validación",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Falta identidad (token inválido o X-Client-Id ausente)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{id}": {
      "get": {
        "tags": [
          "sessions"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "clientIdLegacy": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "as_athlete_id",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Detalle de una sesión",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionDetail"
                }
              }
            }
          },
          "400": {
            "description": "Error de validación",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Falta identidad (token inválido o X-Client-Id ausente)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/cardio-sessions": {
      "post": {
        "tags": [
          "cardio-sessions"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "clientIdLegacy": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "date": {
                    "type": "string",
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                  },
                  "activity_type": {
                    "type": "string",
                    "enum": [
                      "cardio",
                      "tecnico_tactico",
                      "otro"
                    ]
                  },
                  "duration_min": {
                    "type": "integer",
                    "minimum": 0,
                    "exclusiveMinimum": true
                  },
                  "notes": {
                    "type": "string",
                    "maxLength": 500
                  }
                },
                "required": [
                  "date",
                  "activity_type",
                  "duration_min"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Sesión de cardio guardada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardioSession"
                }
              }
            }
          },
          "400": {
            "description": "Error de validación",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Falta identidad (token inválido o X-Client-Id ausente)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "cardio-sessions"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "clientIdLegacy": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "as_athlete_id",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Historial de sesiones de cardio, más reciente primero",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CardioSession"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error de validación",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Falta identidad (token inválido o X-Client-Id ausente)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  }
}